scripts/requirements.txt
pycryptodome
dnslib
